+++ RecipeManager/PieceCommands.cs
@@ -47,13 +47,13 @@ namespace RecipeManager
 
         public override void Run(string[] args)
         {
-            if (Config.EnableDebugMode.Value) { Logger.LogInfo("Starting to dump recipes"); }
-            String ODBRecipes_path = Path.Combine(BepInEx.Paths.ConfigPath, "RecipeManager", "AllPiecesFound.yaml");
-            RecipeModificationCollection recipeCollection = new RecipeModificationCollection();
+            if (Config.EnableDebugMode.Value) { Logger.LogInfo("Starting to dump piece list"); }
+            String ODBRecipes_path = Path.Combine(BepInEx.Paths.ConfigPath, "RecipeManager", "AllPiecesDebug.yaml");
+            PieceModificationCollection recipeCollection = new PieceModificationCollection();
             using (StreamWriter writetext = new StreamWriter(ODBRecipes_path))
             {
                 if (Config.EnableDebugMode.Value) { Logger.LogInfo("Gathering Pieces"); }
-                List<Piece> pieces = Resources.FindObjectsOfTypeAll<Piece>().ToList<Piece>();
+                List<Piece> pieces = Resources.FindObjectsOfTypeAll<Piece>().Where(pc => pc.name.EndsWith("(Clone)") != true).ToList<Piece>();
 
                 foreach (Piece pc in pieces)
                 {
@@ -64,74 +64,84 @@ namespace RecipeManager
                     if (Config.EnableDebugMode.Value) { Logger.LogInfo($"Building Recipe {pc.m_name}"); }
                     PieceModification piece_as_mod = new PieceModification();
 
-                    recipe_as_mod.recipeName = recipe.name;
-                    recipe_as_mod.minStationLevel = (short)recipe.m_minStationLevel;
-                    recipe_as_mod.craftAmount = (short)recipe.m_amount;
-                    recipe_as_mod.action = DataObjects.Action.Enable;
+                    piece_as_mod.action = PieceAction.Enable;
+                    //Logger.LogInfo("Set action");
+                    piece_as_mod.EnablePiece = pc.enabled;
+                    //Logger.LogInfo("Set enable status");
+                    piece_as_mod.CanBeDeconstructed = pc.m_canBeRemoved;
+                    //Logger.LogInfo("Set can be deconstructed");
+                    piece_as_mod.CultivatedGroundOnly = pc.m_cultivatedGroundOnly;
+                    //Logger.LogInfo("Set cultivator ground only");
+                    piece_as_mod.ComfortGroup = pc.m_comfortGroup;
+                    //Logger.LogInfo("Set comfort group");
+                    piece_as_mod.ComfortAmount = pc.m_comfort;
+                    //Logger.LogInfo("Set comfort amount");
+                    piece_as_mod.GroundPlacement = pc.m_groundPiece;
+                    //Logger.LogInfo("Set is ground piece");
+                    piece_as_mod.SpaceRequired = pc.m_spaceRequirement;
+                    //Logger.LogInfo("Set space required");
+                    piece_as_mod.AllowedInDungeon = pc.m_allowedInDungeons;
+                    //Logger.LogInfo("Set allowed in dungeon");
+                    piece_as_mod.CraftingStationConnectionRadius = pc.m_connectRadius;
+                    //Logger.LogInfo("Set crafting radius");
+                    piece_as_mod.MustBeAvobeConnectedStation = pc.m_mustBeAboveConnected;
+                    //Logger.LogInfo("Set must be above connected station");
+                    piece_as_mod.OnlyInSelectBiome = pc.m_onlyInBiome;
+                    //Logger.LogInfo("Set only in biome");
+                    piece_as_mod.PieceCategory = pc.m_category;
+                    //Logger.LogInfo("Set set category");
+                    piece_as_mod.PieceDescription = pc.m_description;
+                    //Logger.LogInfo("Set description");
+                    piece_as_mod.PieceName = pc.m_name;
+                    //Logger.LogInfo("Set piecename");
+                    piece_as_mod.prefab = pc.gameObject.name;
+                    //Logger.LogInfo("Set prefabname");
+                    if (pc.m_craftingStation != null) {
+                        piece_as_mod.RequiredToPlaceCraftingStation = pc.m_craftingStation.name;
+                        //Logger.LogInfo("Set set crafting station");
+                    }
+                    piece_as_mod.IsUpgradeForStation = pc.m_isUpgrade;
+                    //Logger.LogInfo("Set is upgrade");
+                    // Build the simple requirements list
 
-                    // null checks needed on all child object accessors due to some oddly formed recipes
-                    if (Config.EnableDebugMode.Value) { Logger.LogInfo("Checking for empty referenced objects"); }
-                    if (recipe.m_craftingStation != null)
+                    List<SimpleRequirement> requirements = new List<SimpleRequirement>() { };
+                    if (pc.m_resources != null && pc.m_resources.Length > 0)
                     {
-                        if (Config.EnableDebugMode.Value) { Logger.LogInfo("Adding crafting station"); }
-                        recipe_as_mod.craftedAt = recipe.m_craftingStation.name;
-                    }
-                    if (recipe.m_item != null)
-                    {
-                        if (Config.EnableDebugMode.Value) { Logger.LogInfo("Adding prefab"); }
-                        recipe_as_mod.prefab = recipe.m_item.name;
-                    }
-                    SimpleRecipe temp_srecipe = new SimpleRecipe();
-                    if (recipe.m_resources != null && recipe.m_resources.Length > 0)
-                    {
-                        temp_srecipe.anyOneResource = recipe.m_requireOnlyOneIngredient;
-                        if (Config.EnableDebugMode.Value) { Logger.LogInfo("Building resource requirements"); }
-                        foreach (Piece.Requirement req in recipe.m_resources)
+                        // if (Config.EnableDebugMode.Value) { Logger.LogInfo("Building resource requirements"); }
+                        foreach (Piece.Requirement req in pc.m_resources)
                         {
                             try
                             {
-                                Ingrediant res_req = new Ingrediant();
-                                if (Config.EnableDebugMode.Value) { Logger.LogInfo("Setting crafting cost"); }
-                                res_req.craftCost = (short)req.m_amount;
-                                if (Config.EnableDebugMode.Value) { Logger.LogInfo("Setting upgrade cost"); }
-                                res_req.upgradeCost = (short)req.m_amountPerLevel;
-                                if (Config.EnableDebugMode.Value) { Logger.LogInfo("Setting refund bool"); }
-                                res_req.refund = req.m_recover;
-                                if (Config.EnableDebugMode.Value) { Logger.LogInfo("Setting prefab name"); }
-                                if (req.m_resItem != null) { res_req.prefab = req.m_resItem.name; }
-                                if (Config.EnableDebugMode.Value) { Logger.LogInfo("Adding to ingrediants list"); }
-                                temp_srecipe.ingredients.Add(res_req);
+                                requirements.Add(new SimpleRequirement() { amount = req.m_amount, Prefab = req.m_resItem.name });
                             }
                             catch (Exception e)
                             {
-                                if (Config.EnableDebugMode.Value) { Logger.LogWarning($"Requirement did not contain all of the details required to set an ingrediant {req} \n{e}"); }
+                                if (Config.EnableDebugMode.Value) { Logger.LogWarning($"Piece requirement setup error {req} \n{e}"); }
                             }
                         }
                     }
-                    if (temp_srecipe.ingredients != null && temp_srecipe.ingredients.Count > 0)
-                    {
-                        recipe_as_mod.recipe = temp_srecipe;
-                    }
-                    if (Config.EnableDebugMode.Value) { Logger.LogInfo($"Adding {recipe.name} to collection."); }
-                    if (recipeCollection.RecipeModifications.ContainsKey(recipe.name))
+                    piece_as_mod.requirements = requirements;
+
+                    if (Config.EnableDebugMode.Value) { Logger.LogInfo($"Adding {piece_as_mod.PieceName} to collection."); }
+                    if (recipeCollection.PieceModifications.ContainsKey(piece_as_mod.PieceName))
                     {
                         try
                         {
                             int randtag = UnityEngine.Random.Range(0, 1000);
-                            recipeCollection.RecipeModifications.Add(recipe.name + $"_{randtag}", recipe_as_mod);
-                            Logger.LogWarning($"{recipe.name} was already added to the list of recipes and will be renamed {recipe.name}_{randtag}, please use unique recipe names.");
+                            recipeCollection.PieceModifications.Add(piece_as_mod.PieceName + $"_{randtag}", piece_as_mod);
+                            Logger.LogWarning($"{piece_as_mod.PieceName} was already added to the list of recipes and will be renamed {piece_as_mod.PieceName}_{randtag}, please use unique piece modification names.");
                         }
                         catch
                         {
-                            Logger.LogWarning($"{recipe.name} was already added to the list of recipes and will be skipped, please use unique recipe names.");
+                            Logger.LogWarning($"{piece_as_mod.PieceName} was already added to the list of recipes and will be skipped, please use unique recipe names.");
                         }
                     }
                     else
                     {
-                        recipeCollection.RecipeModifications.Add(recipe.name, recipe_as_mod);
+                        recipeCollection.PieceModifications.Add(piece_as_mod.PieceName, piece_as_mod);
                     }
 
-                    if (Config.EnableDebugMode.Value) { Logger.LogInfo($"Recipe {recipe.name} Added."); }
+                    if (Config.EnableDebugMode.Value) { Logger.LogInfo($"Piece {piece_as_mod.PieceName} Added."); }
                 }
                 if (Config.EnableDebugMode.Value) { Logger.LogInfo($"Serializing and printing recipes."); }
                 var yaml = Config.yamlserializer.Serialize(recipeCollection);
@@ -140,3 +150,4 @@ namespace RecipeManager
             }
         }
     }
+}