+++ RecipeManager/Package/CHANGELOG.md
@@ -1,3 +1,14 @@
+ **0.4.0**
+---
+```
+- Add support for piece modifications
+	- Details about pieces can now be modified
+	- Piece modifications are seperated from item modifications
+- Added support for multiple config files of each type
+- Unified synchronization for multiple files & piece or recipe types
+- Virtualized piece and recipe synchronization, files are no longer created or mirrored from the host/server
+```
+
  **0.3.4**
 ---
 ```